webapp/requirements/base.txt
Django==2.2.3
djangorestframework==3.9.4
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
